Move out syntax node analysis implementations out of CodeAnalayzerManager
I am planning to use the same code to run sysntax node analysis in code generators
